% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{block.sizes.4group}
\alias{block.sizes.4group}
\title{given parameter values, compute block sizes}
\usage{
block.sizes.4group(sim.settings)
}
\arguments{
\item{sim.settings}{a list whose entries contain the parameter values
which govern this random graph draw (see details)}
}
\value{
a named vector; names are block names, values are block sizes
}
\description{
given the parameter values N, p.H, and p.F, compute the number of vertices
in each of the four blocks: FnotH, FH, notFnotH, notFH
}

