% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{draw.4group.graph}
\alias{draw.4group.graph}
\title{draw a random graph from a stochastic blockmodel with four blocks}
\usage{
draw.4group.graph(sim.settings, type = "simple")
}
\arguments{
\item{sim.settings}{a list whose entries contain the parameter values
which govern this random graph draw (see details)}

\item{type}{either 'simple' or 'nested', depending on the type of inhomogenous
mixing; see \code{pref.matrix.4group}}
}
\value{
an igraph graph, drawn according to the
        settings passed in
}
\description{
Draw a random graph from a stochastic blockmodel where the groups are
determined by membership in the frame population (F) and membership in
the hidden population (H). Membership in F is assumed to be
independent of membership in H, and vice-versa.
}
\details{
The four groups, as determined by membership in F and H, are
FH, FnotH, notF,notH, notFH.

The \code{sim.settings} list should have entries
\itemize{
\item N the size of the entire population
\item p.F the prevalence of the frame population (N.F/N)
\item p.H the prevalence of the hidden population (N.H/N)
\item pi.within the probability of forming a tie for a pair of
      nodes in the same group
\item rho the relative probability of edge between two nodes that
      are not in the same group (for example, between a node in FnotH and FH).
      if rho is 0.8, then edges between FnotH and FH are 80% as likely as edges between
      FnotH and FnotH (or FH and FH)
\item tau the true positive rate
}

The overall level of connectivity in the graph can be controlled by
varying pi.within. The extent to which members of the group are
homogenously mixed can be controlled by varying rho.
For example, when rho is 1, then all four groups are perfectly
mixed with the entire population. When rho is less than 1, members of the
same group tend to form ties with one another more than with others.
And when rho is greater than 1, members of the same group tend to
form ties with others more than with one another.
}

