% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scratch.R
\name{draw.FH.graph}
\alias{draw.FH.graph}
\title{draw.FH.graph}
\usage{
draw.FH.graph(sim.settings)
}
\arguments{
\item{sim.settings}{a list whose entries contain the parameter values
which govern this random graph draw (see details)}
}
\value{
an igraph graph, drawn according to the
        settings passed in
}
\description{
Draw a random graph from a stochastic blockmodel where the groups are
determined by membership in the frame population (F) and membership in
the hidden population (H). Membership in F is assumed to be
independent of membership in H, and vice-versa.
}
\details{
The four groups, as determined by membership in F and H, are
FH, FnotH, notF,notH, notFH.

The \code{sim.settings} list should have entries
\itemize{
\item N the size of the entire population
\item p.F the prevalence of the frame population (N.F/N)
\item p.H the prevalence of the hidden population (N.H/N)
\item pi.within the probability of forming a tie for a pair of
      nodes in the same group
\item rho.FnotF the relative probability of edge between a node in F
      and a node not in F (pi.FnotF = pi.within * rho). so if this is
      0.8, then edges between F and not F are 80% as likely as edges between
      F and F (or not F and not F)
\item rho.HnotH the relative probability of an edge between a node in H
      and a node not in H (see rho.FnotF)
\item tau the true positive rate
}

The overall level of connectivity in the graph can be controlled by
varying pi.within. The extent to which members of the group are
homogenously mixed can be controlled by varying rho.FnotF and rho.HnotH.
For example, when rho.FnotF is 1, then the frame population is perfectly
mixed with the entire population. When it is less than 1, members of the
frame population tend to form ties with one another more than with others.
And when it is greater than 1, members of the frame population tend to
form ties with others more than with one another.
}

