% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{draw.sbm.graph}
\alias{draw.sbm.graph}
\title{draw a random graph from a stochastic blockmodel and label the vertices}
\usage{
draw.sbm.graph(block.sizes, pref.matrix, block.names = NULL)
}
\arguments{
\item{block.sizes}{a vector with the number of vertices in each block}

\item{pref.matrix}{a matrix whose (i,j)th entry is the probability of
an edge between a vertex in the block containing i and a vertex
in the block containing j}

\item{block.names}{if not NULL, then names of the blocks}
}
\value{
the \code{igraph} object with each vertex's group membership
        given by a vertex attribute
}
\description{
Draw a random graph from a stochastic blockmodel using a k-vector
of block sizes and a k x k preference matrix
}
\details{
The \code{igraph} object that gets returned is a graph drawn from
the stochastic block-model. The vertices each have a 'group' attribute
containing the name of the block they have been assigned to. These block
names default to 'group.1', 'group.2', ... if none are specified.
}

