% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{ev.4group}
\alias{ev.4group}
\title{compute the expected value for various quantities under four-group model}
\usage{
ev.4group(these.params, type = "nested", inF = c(1, 1, 0, 0), inH = c(0,
  1, 0, 1), inU = c(1, 1, 1, 1))
}
\arguments{
\item{these.params}{list containing the simulation parameters}

\item{type}{either "nested" or "simple"}

\item{inF}{vector with 1s for blocks in F, 0 otherwise}

\item{inH}{vector with 1s for blocks in H, 0 otherwise}

\item{inU}{vector with 1s for blocks in U, 0 otherwise}
}
\value{
the expected value of various quantities
}
\description{
compute the expected value for various quantities under four-group model
}

