% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{ev.4group.tot}
\alias{ev.4group.mean}
\alias{ev.4group.tot}
\title{compute the expected value for various quantities under four-group model}
\usage{
ev.4group.tot(pm, bs, fg, sg)

ev.4group.mean(pm, bs, fg, sg)
}
\arguments{
\item{pm}{the preference matrix}

\item{bs}{the block sizes}

\item{fg}{0/1 indicator vector for blocks in the first index (A in the description above)}

\item{sg}{0/1 indicator vector for blocks in the second index (B in the description above)}
}
\value{
the expected value of the mean or the total number of connections
        from the first group to the second. Means are taken with respect to the members
        of the first index
}
\description{
for groups A and B, these functions return \eqn{d_{A,B}} and \eqn{\bar{d}_{A,B}}
}

