% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existing_igraph.R
\name{existing_params}
\alias{existing_params}
\title{create a parameter object for an existing igraph}
\usage{
existing_params(params)
}
\arguments{
\item{params}{the parameters}
}
\value{
an object with the parameters
}
\description{
Given an existing igraph object, add attributes for the frame
population and for the hidden population.
}
\details{
The \code{params} vector should have at least three entries:
\itemize{
\item \code{groups} is a vector that has a list of attributes that
each node in the \code{igraph} object \code{g} has. These attributes
should all have values of 0/1 or TRUE/FALSE.
\item \code{gps.in.F} is a vector that has a list of group names; nodes in
those groups are taken to be in the frame population
\item \code{gps.in.H} is a vector that has a list of group names; nodes in
those groups are taken to be in the hidden population
}'
}
\examples{
my_params <- existing_params(list(group.names=c('a', 'b', 'c', 'd'),
                                  gps.in.F=c('b', 'c'),
                                  gps.in.H=c('a', 'b')))
}
