% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{expected.edgecount.4group}
\alias{expected.edgecount.4group}
\title{compute the expected number of connections (degree) between two groups under 4-block model}
\usage{
expected.edgecount.4group(pref.matrix, block.sizes, first.group, second.group)
}
\arguments{
\item{pref.matrix}{the preference matrix}

\item{block.sizes}{vector with the number of vertices in each block}

\item{first.group}{vector with 0/1 entries indicating which blocks are in group A}

\item{second.group}{vector with 0/1 entries indicating which blocks are in group B}
}
\value{
The expected number of edges between the two groups under the stochastic block
        model given by pref.matrix. For example,
}
\description{
for groups A and B, this returns \eqn{d_{A,B} = d_{B,A}}.
}

