% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{expected_edgecount}
\alias{expected_edgecount}
\title{generic function for getting expected edgecounts}
\usage{
expected_edgecount(params, ...)
}
\arguments{
\item{params}{parameter vector describing the model (the class of this param vector
determines which method gets run)}
}
\description{
generic function for getting expected edgecounts
}
