% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{expected_edgecount.sbm_4group}
\alias{expected_edgecount.sbm_4group}
\title{compute the expected number of connections (degree) between two groups under 4-block model}
\usage{
\method{expected_edgecount}{sbm_4group}(params, first.group, second.group)
}
\arguments{
\item{params}{a parameter vector, which must have pref.matrix and block.sizes}

\item{first.group}{vector with the names of the block that blocks that are in group A}

\item{second.group}{vector with the names of the blocks that are in group B}
}
\value{
The expected number of edges between the two groups under the stochastic block 
        model given by pref.matrix. For example,
}
\description{
for groups A and B, this returns \eqn{d_{A,B} = d_{B,A}}.
}
\details{
it does not matter how the preference matrix and block sizes were generated,
so this function could be used with many different models.
}

