% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_graph.R
\name{frame_census}
\alias{frame_census}
\title{Create a set of parameters describing a census of F in graph}
\usage{
frame_census(params = NULL)
}
\arguments{
\item{params}{a list of params, which must contain tau}
}
\value{
a reporting parameter object
}
\description{
This creates an object of class 'census_F'
}
