% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample_graph.R
\name{frame_srs}
\alias{frame_srs}
\title{Create a set of parameters describing a simple random sample of F in graph}
\usage{
frame_srs(params)
}
\arguments{
\item{params}{a list of params, which must contain sampling.frac, the sampling fraction}
}
\value{
a reporting parameter object
}
\description{
This creates an object of class 'srs_F', which has a simple
random sample without replacement of the frame population nodes
(NB: Sarndal et al call this SI sampling)
}

