% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existing_igraph.R
\name{generate_graph.existing_igraph}
\alias{generate_graph.existing_igraph}
\title{create a social network from an existing igraph object}
\usage{
\method{generate_graph}{existing_igraph}(params, g)
}
\arguments{
\item{params}{an optional list containing hidden.pop and frame.pop}
}
\value{
the \code{igraph} object with each vertex's hidden and frame pop membership 
indicated by in.F and in.H
}
\description{
Given an existing igraph object, add attributes for the frame
population and for the hidden population.
}
