% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generate_graph}
\alias{generate_graph}
\title{generate a social network}
\usage{
generate_graph(params, ...)
}
\arguments{
\item{params}{object with the parameters governing how the network
should be generated}
}
\value{
the \code{igraph} object with the generated social network
}
\description{
this is a generic function to generate a social network
}

