% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{generate_graph.sbm}
\alias{generate_graph.sbm}
\title{draw a random graph from a 4group_oneparam_simple stochastic blockmodel}
\usage{
\method{generate_graph}{sbm}(params)
}
\arguments{
\item{params}{a list containing block.sizes, pref.matrix, and optionally block.names}
}
\value{
the \code{igraph} object with each vertex's group membership 
        given by a vertex attribute
}
\description{
Draw a random graph from a stochastic blockmodel using a k-vector
of block sizes and a k x k preference matrix
}
\details{
This routine is agnostic about the algorithm that was used to produce
the preference matrix and the block sizes, so it could potentially be
used with a lot of different models.

The \code{igraph} object that gets returned is a graph drawn from
the stochastic block-model. The vertices each have a 'group' attribute
containing the name of the block they have been assigned to. These block
names default to 'group.1', 'group.2', ... if none are specified.
}

