% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample_graph.R
\name{hidden_relprob}
\alias{hidden_relprob}
\title{Create a set of parameters describing a relative probability sample of H in graph}
\usage{
hidden_relprob(params)
}
\arguments{
\item{params}{a list of params, which must contain sampling.frac, the sampling fraction,
and inclusion.trait, the trait to use for relative probabilities (typically degree)}
}
\value{
a reporting parameter object
}
\description{
This creates an object of class 'relprobsample_H', which has a relative
probability sample of H without replacement
}

