% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reporting_graph.R
\name{imperfect_reporting}
\alias{imperfect_reporting}
\title{Create a set of parameters describing reporting behavior that is not perfectly accurate}
\usage{
imperfect_reporting(params)
}
\arguments{
\item{params}{a list of params, which must contain tau}
}
\value{
a reporting parameter object
}
\description{
Create a set of parameters describing reporting behavior that is not perfectly accurate
}
\details{
For now, the only simulation parameter that makes a difference is
\eqn{\tau}.
}

