% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{label.vertices}
\alias{label.vertices}
\title{assign labels to vertices in a graph}
\usage{
label.vertices(graph, block.sizes, block.names)
}
\arguments{
\item{graph}{the \code{igraph} graph}

\item{block.sizes}{a vector with the number of nodes in each group}

\item{block.names}{a vector (same length as \code{block.sizes}) with names of each group}
}
\description{
assign labels to vertices in a graph
}

