% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrsimulatr_package.R
\docType{package}
\name{nrsimulatr}
\alias{nrsimulatr}
\title{network reporting simulations}
\description{
Utilities for drawing random social networks, reporting networks,
and network reporting data using different assumptions about the process
that generates connections between the nodes. Useful for conducting simulation
studies to learn about sampling properties of network reporting and other
network sampling methods.
}
