% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reporting_graph.R
\name{perfect_reporting}
\alias{perfect_reporting}
\title{Create a set of parameters describing reporting behavior that is not perfectly accurate}
\usage{
perfect_reporting(params = NULL)
}
\arguments{
\item{params}{a list of params, which must contain tau}
}
\value{
a reporting parameter object
}
\description{
This creates an object of class 'imperfect_reporting' whose parameter values
correspond to perfect reporting.
}
\details{
For now, the only simulation parameter that makes a difference is
\eqn{\tau}.
}

