% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{pref.matrix.4group}
\alias{pref.matrix.4group}
\title{construct a 4-block preference matrix based on two independent categories}
\usage{
pref.matrix.4group(block.sizes, pi.within, rho, type, inF = NULL,
  inH = NULL)
}
\arguments{
\item{pi.within}{the probability of an edge between two people in the same group}

\item{rho}{the relative probability of an edge between two people who are not
in the same group (see Description)}

\item{type}{(see Description)}
}
\value{
a 4x4 preference matrix
}
\description{
TODO
}

