% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{report.sbm.edges}
\alias{report.sbm.edges}
\title{for each vertex, compute the number of reported connections to each block}
\usage{
report.sbm.edges(g, prefix = "d.", mode = "all")
}
\arguments{
\item{g}{the \code{igraph} object}

\item{prefix}{the prefix to use in the variable names that are attached
(useful if this function will be used to compute reports more than once)}

\item{mode}{one of "all", "in", or "out". see Details}
}
\value{
the \code{igraph} object with new attributes affixed to each vertex
        (see Details)
}
\description{
for each vertex, compute the number of reported connections to each block
}
\details{
This function is useful for computing edge counts (ie, reports) between
individual vertices and the blocks. These edge counts are the building
blocks of many network reporting estimators.

Note that, counter-intuitively, mode="in" will compute out-reports and
mode="out" will compute in-reports. This is the \code{igraph} convention.
For undirected graphs (for example, when computing degrees in the social
network), use "all".

For example, if the prefix is "y." and the mode is "in", we are asking
\code{report.sbm.edges} to count each vertex's number of out-reports
about each block. If we have three blocks, A, B, and C, then in the
graph that \code{report.sbm.edges} returns, each vertex will have
four new attributes: \code{y.degree}, \code{y.A}, \code{y.B}, and \code{y.C}.
}

