% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{report_detailed_edges}
\alias{report_detailed_edges}
\title{report detailed edges}
\usage{
report_detailed_edges(g, ...)
}
\arguments{
\item{g}{the social network or reporting graph}
}
\value{
the \code{igraph} object with the generated social network
}
\description{
this is a generic function to report edges one by one from a social network 
or a reporting graph
}
