% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{report_edges}
\alias{report_edges}
\title{report edges}
\usage{
report_edges(g, ...)
}
\arguments{
\item{g}{the social network or reporting graph}
}
\value{
the \code{igraph} object with the generated social network
}
\description{
this is a generic function to report edges from a social network or
a reporting graph
}
