% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{report_edges.sbm}
\alias{report_edges.sbm}
\title{for each vertex, compute the number of  connections to each block}
\usage{
\method{report_edges}{sbm}(g, prefix = "d.", mode = "all")
}
\arguments{
\item{g}{the \code{igraph} object}

\item{prefix}{the prefix to use in the variable names that are attached
(useful if this function will be used to compute reports more than once)}

\item{mode}{one of "all", "in", or "out". see Details}
}
\value{
the \code{igraph} object with new attributes affixed to each vertex
        (see Details)
}
\description{
If this function is called on an undirected social network graph (with mode='all'),
then it counts the true number of social connections between each node and
the various blocks.
}
\details{
On the other hand, if this function is called on a directed reporting graph,
then it computes the reported number of social connections (with mode='in')
or visibility (with mode='out').



This function is useful for computing edge counts (i.e. )
between
individual vertices and the blocks. These edge counts are the building
blocks of many network reporting estimators.

Since this function is based on the stochastic block model, it assumes
that the groups are mutually exclusive. This could be modified in the
future.

Note that, counter-intuitively, mode="in" will compute out-reports and
mode="out" will compute in-reports. This is the \code{igraph} convention.
For undirected graphs (for example, when computing degrees in the social
network), use "all".

For example, if the prefix is "y." and the mode is "in", we are asking
\code{report.sbm.edges} to count each vertex's number of out-reports
about each block. If we have three blocks, A, B, and C, then in the
graph that \code{report.sbm.edges} returns, each vertex will have
four new attributes: \code{y.degree}, \code{y.A}, \code{y.B}, and \code{y.C}.
}
