% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TO_REMOVE_reporting_estimates.R
\name{reporting.estimates}
\alias{reporting.estimates}
\title{compute network reporting estimates from a randomly simulated graph}
\usage{
reporting.estimates(reporting.graph, hidden.popn, frame.popn)
}
\arguments{
\item{reporting.graph}{the reporting graph object}

\item{hidden.popn}{the names of block(s) containing the hidden population}

\item{frame.popn}{the names of block(s) containing the frame population}
}
\value{
a data frame with estimates and several related quantities
}
\description{
compute network reporting estimates from a randomly simulated graph
}

