% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reporting_graph.R
\name{reporting.graph}
\alias{reporting.graph}
\title{Create a reporting graph from a social network}
\usage{
reporting.graph(sim.graph, hidden.popn, frame.popn)
}
\arguments{
\item{sim.graph}{the \code{igraph} object with the social network}

\item{hidden.popn}{the name(s) of blocks containing hidden population members}

\item{frame.popn}{the name(s) of blocks containing frame population members}
}
\value{
the \code{igraph} object with each vertex's group membership
        given by vertex label
}
\description{
Note that this function relies upon the fact that the igraph object
\code{sim.graph} will have an attribute called \code{'sim.settings'},
which is a list with parameters describing the simulation.
}
\details{
For now, the only simulation parameter that makes a difference is
\eqn{\tau}.
}

