% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_graph.R
\name{reporting_graph.imperfect_reporting}
\alias{reporting_graph.imperfect_reporting}
\title{Create a reporting graph from a social network}
\usage{
\method{reporting_graph}{imperfect_reporting}(reporting.params, sim.graph)
}
\arguments{
\item{reporting.params}{the reporting parameters}

\item{sim.graph}{the \code{igraph} object with the social network}
}
\value{
the \code{igraph} object for the directed reporting graph
}
\description{
Note that this function relies upon the fact that the igraph object
\code{sim.graph} will have an attribute called \code{'sim.settings'},
which is a list with parameters describing the simulation.
}
\details{
For now, the only simulation parameter that makes a difference is
\eqn{\tau}.
}

