% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_graph_ind.R
\name{reporting_graph.imperfect_reporting_ind}
\alias{reporting_graph.imperfect_reporting_ind}
\title{Create a reporting graph from a social network}
\usage{
\method{reporting_graph}{imperfect_reporting_ind}(reporting.params, sim.graph, stochastic = FALSE)
}
\arguments{
\item{reporting.params}{the reporting parameters}

\item{sim.graph}{the \code{igraph} object with the social network}

\item{stochastic}{if TRUE, then treat the reporting parameters as expected 
values from bernoulli trials for whether or not each edge is observed; otherwise,
reporting params are deterministic (but since edges are discrete, this could lead to
rounding issues -- ie, tau=0.8 for 3 edges would produce 2 edges observed)}
}
\value{
the \code{igraph} object for the directed reporting graph
}
\description{
Create a reporting graph with attributes added to vertices in the network
that have aggregate relational reports. Incorporate imperfect reporting
(i.e., false positives and negatives) at the individual level (varying
from person to person).
}
\details{
Note that this function relies upon the fact that the igraph object
\code{sim.graph} will have an attribute called \code{'sim.settings'},
which is a list with parameters describing the simulation.


For now, the only simulation parameter that makes a difference is
\eqn{\tau}. For this individual-level imperfect reporting, the
entry in the \code{reporting.params} list called \code{tau} should
be a function that takes two arguments: a vertex id and a graph.
It should return a value from 0 to 1, which is the tau for reports
from the given vertex.
}
