% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{reporting_graph}
\alias{reporting_graph}
\title{make a reporting graph from a social network}
\usage{
reporting_graph(params, g, ...)
}
\arguments{
\item{params}{object with the parameters governing reporting (i.e., how the
reporting network should be generated}
}
\value{
the \code{igraph} object with the generated social network
}
\description{
this is a generic function to generate a reporting graph from
reporting parameters and a social network
}

