% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_graph.R
\name{reporting_graph_detailed.imperfect_reporting}
\alias{reporting_graph_detailed.imperfect_reporting}
\title{Create a detailed reporting graph from a social network (i.e., not ARD)}
\usage{
\method{reporting_graph_detailed}{imperfect_reporting}(reporting.params, sim.graph)
}
\arguments{
\item{reporting.params}{the reporting parameters}

\item{sim.graph}{the \code{igraph} object with the social network}
}
\value{
the \code{tibble} object with rows for the directed reporting graph
}
\description{
Create a reporting graph with attributes added to vertices in the network
that have detailed individual reports (rather than ARD). 
Incorporate imperfect reporting
(i.e., false positives and negatives) at the aggregate level.
}
\details{
Note that this function relies upon the fact that the igraph object
\code{sim.graph} will have an attribute called \code{'sim.settings'},
which is a list with parameters describing the simulation.


For now, the only simulation parameter that makes a difference is
\eqn{\tau}.
}
