% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{reporting_graph_detailed}
\alias{reporting_graph_detailed}
\title{make a detailed reporting graph from a social network}
\usage{
reporting_graph_detailed(params, g, ...)
}
\arguments{
\item{params}{object with the parameters governing reporting (i.e., how the 
reporting network should be generated}
}
\value{
the \code{igraph} object with the generated social network
}
\description{
this is a generic function to generate a reporting graph with
detailed reports from
reporting parameters and a social network
}
