% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample_graph.R
\name{sample_graph.census_H}
\alias{sample_graph.census_H}
\title{Get dataset from a census of H in reporting graph}
\usage{
\method{sample_graph}{census_H}(params, reporting.graph)
}
\arguments{
\item{params}{sampling parameters; in this case, anything of the class 'census_H';
its actual contents are ignored}

\item{reporting.graph}{the reporting graph}
}
\value{
a dataset made from all of the nodes in H
}
\description{
Get dataset from a census of H in reporting graph
}

