% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_graph.R
\name{sample_graph.census_U}
\alias{sample_graph.census_U}
\title{Get dataset from a census of U in reporting graph}
\usage{
\method{sample_graph}{census_U}(params, reporting.graph)
}
\arguments{
\item{params}{sampling parameters; in this case, anything of the class 'census_U';
its actual contents are ignored}

\item{reporting.graph}{the reporting graph}
}
\value{
a dataset made from all of the nodes in U
}
\description{
Get dataset from a census of U in reporting graph
}
