% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sample_graph}
\alias{sample_graph}
\title{produce dataset from a sample of a reporting graph or social network}
\usage{
sample_graph(params, g, ...)
}
\arguments{
\item{params}{object with the parameters governing how the sample should
be obtained}

\item{g}{the social network or reporting graph to sample'}
}
\value{
dataframe with the attributes of the sampled vertices
}
\description{
this is a generic function to generate a dataset with information
from a sample of nodes in a reporting graph or social network.
the parameters describe how the sample is obtained.
}

