% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_graph.R
\name{sample_graph.relprob_H}
\alias{sample_graph.relprob_H}
\title{Get dataset from a relative probability sample of H in reporting graph}
\usage{
\method{sample_graph}{relprob_H}(params, reporting.graph)
}
\arguments{
\item{params}{sampling parameters; see hidden_relprob}

\item{reporting.graph}{the reporting graph}
}
\value{
a dataset made from a simple random sample of the nodes in F
}
\description{
Get dataset from a relative probability sample of H in reporting graph
}

