% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample_graph.R
\name{sample_graph.srs_F}
\alias{sample_graph.srs_F}
\title{Get dataset from a simple random sample of F in reporting graph}
\usage{
\method{sample_graph}{srs_F}(params, reporting.graph)
}
\arguments{
\item{params}{sampling parameters; see frame_srs}

\item{reporting.graph}{the reporting graph}
}
\value{
a dataset made from a simple random sample of the nodes in F
}
\description{
Get dataset from a simple random sample of F in reporting graph
}

