% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_block_sizes}
\alias{sbm_block_sizes}
\title{given parameter values for a stochastic block model, compute block sizes (generic)}
\usage{
sbm_block_sizes(params)
}
\arguments{
\item{params}{a list whose entries contain the parameter values
which govern this random graph draw (see details)}
}
\value{
a named vector; names are block names, values are block sizes
}
\description{
given the parameter values of a stochastic block model, return a named
list of the block sizes
}
