% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_block_sizes.sbm_4group}
\alias{sbm_block_sizes.sbm_4group}
\title{given parameter values, compute block sizes}
\usage{
\method{sbm_block_sizes}{sbm_4group}(params)
}
\arguments{
\item{params}{a list whose entries contain the parameter values
which govern this random graph draw (see details)}
}
\value{
a list; block.sizes has a vector whose names are block names, values are block sizes;
        gps.in.F is a vector with the names of groups whose members are in F; and
        gps.in.H is a vector with the names of groups whose members are in H
}
\description{
given the parameter values N, p.H, and p.F, compute the number of vertices
in each of the four blocks: FnotH, FH, notFnotH, notFH
}

