% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_estimates}
\alias{sbm_estimates}
\title{compute network reporting estimates from a randomly simulated stochastic block model}
\usage{
sbm_estimates(data, weights = data[, "sampling.weight"])
}
\arguments{
\item{weights}{weights that account for the sampling (if any);
this should be a vector that is the same length as the number of rows in rg.dat}

\item{rg.dat}{dataset from a reporting graph object}
}
\value{
a data frame with estimates and several related quantities
}
\description{
note that this code relies upon the fact that the vertices should have
two 0/1 attributes, in.H and in.F, which indicate whether or not each one
is in the hidden population and the frame population
}

