% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_ev}
\alias{sbm_ev}
\title{generic fn for getting expected values from stochastic block model}
\usage{
sbm_ev(params, ...)
}
\arguments{
\item{params}{the parameter object}
}
\description{
generic fn for getting expected values from stochastic block model
}
