% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_ev.sbm_4group}
\alias{sbm_ev.sbm_4group}
\title{compute the expected value for various quantities under four-group model}
\usage{
\method{sbm_ev}{sbm_4group}(
  params,
  inF = c("FnotH", "FH"),
  inH = c("FH", "notFH"),
  inU = c("FnotH", "FH", "notFnotH", "notFH")
)
}
\arguments{
\item{params}{list/object containing the simulation parameters}

\item{inF}{vector with 1s for blocks in F, 0 otherwise}

\item{inH}{vector with 1s for blocks in H, 0 otherwise}

\item{inU}{vector with 1s for blocks in U, 0 otherwise}
}
\value{
the expected value of various quantities
}
\description{
compute the expected value for various quantities under four-group model
}
