% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_label_vertices}
\alias{sbm_label_vertices}
\title{assign labels to vertices in a graph}
\usage{
sbm_label_vertices(graph, block.sizes, block.names)
}
\arguments{
\item{graph}{the \code{igraph} graph}

\item{block.sizes}{a vector with the number of nodes in each group}

\item{block.names}{a vector (same length as \code{block.sizes}) with names of each group}
}
\description{
assign labels to vertices in a graph
}
