% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_params}
\alias{sbm_params}
\title{create a set of simulation parameters for the stochastic block model}
\usage{
sbm_params(params, type)
}
\arguments{
\item{params}{the simulation parameters}

\item{type}{currently, this can be one of '4group_1param_simple', '4group_1param_nested', or '4group_2param'; see the vignette for more info}
}
\value{
an object with the simulation parameters
}
\description{
create a set of simulation parameters for the stochastic block model
}

