% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_pref_matrix.4group_1param_nested}
\alias{sbm_pref_matrix.4group_1param_nested}
\title{construct a 4-block preference matrix based on two independent categories, nested}
\usage{
\method{sbm_pref_matrix}{`4group_1param_nested`}(params)
}
\arguments{
\item{list}{of parameters, including block.sizes, zeta, rho, gps.in.F, and gps.in.H}
}
\value{
a 4x4 preference matrix
}
\description{
TODO
}
