% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_pref_matrix.4group_1param_simple}
\alias{sbm_pref_matrix.4group_1param_simple}
\title{construct a 4-block preference matrix based on two independent categories, simple}
\usage{
\method{sbm_pref_matrix}{4group_1param_simple}(params)
}
\arguments{
\item{list}{of parameters, including block.sizes, zeta and rho}
}
\value{
a 4x4 preference matrix
}
\description{
TODO
}

