% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_pref_matrix.4group_2param}
\alias{sbm_pref_matrix.4group_2param}
\title{construct a 4-block preference matrix based on two independent categories,
and two independent interaction parameters}
\usage{
\method{sbm_pref_matrix}{4group_2param}(params)
}
\arguments{
\item{list}{of parameters, including block.sizes, pi.within, rho, xi, gps.in.F, and gps.in.H}
}
\value{
a 4x4 preference matrix
}
\description{
TODO
}

