% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{sbm_pref_matrix}
\alias{sbm_pref_matrix}
\title{construct a preference matrix for a stochastic block model}
\usage{
sbm_pref_matrix(params)
}
\arguments{
\item{params}{the parameters specific to this type of model}
}
\value{
a preference matrix
}
\description{
TODO
}

